moka miko/requirements.txt
numpy<2.0
faiss-cpu>=1.8.0
torch==2.3.1
transformers==4.41.2
groq==0.9.0
langdetect==1.0.9
sentence-transformers==3.0.1
flask
httpx==0.27.0
